Simplify tests
Apply the same pattern in the run tests for plain_iface to the run tests
for atf_iface: define an auxiliary class to abstract the execution of the
helper test programs, and use it to avoid lots of repetition in the test
cases themselves.

This change was r181 in Subversion.