Plain test programs require a test suite to be define
The original commit for the plain_iface intended for the plain test
programs to include a definition of the test suite they belong to.
However, such change was done late and it did not cover for the changes
to the lua files.  Do this now and add tests.

This change was r182 in Subversion.