Split get_test_program into atf and plain versions
This should have been done from the beginning instead of trying to generalize
the function, but I did not plan well.  The reason is that I now need to add
the ability to define a timeout on a per-test-suite basis, and this requires
independent handling in the plain interface only.

No functional changes.

This change was r183 in Subversion.