Implement timeout parametrization for plain test programs
Add a new 'timeout' field to the definition of plain test programs so that
it is possible to specify an execution deadline for their main test case.
This timeout is optional and defaults to the same value as the timeout for
atf test programs.

This change was r184 in Subversion.