Run the bootstrap test suite as part of the Kyua test suite
Add a new Kyuafile to integrate the bootstrap test suite into regular
runs.  This is to exercise the plain interface in a real use case,
although it does not provide much value by itself.

'make check' runs the bootstrap test suite separately (i.e. outside
of Kyua) to ensure it really yields the expected results and is not
interfered by Kyua.

However, this forces the need to install the 'testsuite' bootstrap
test suite, so this may come in handy for debugging purposes anyway
if we need to tell users to run the test suite by hand.

This change was r186 in Subversion.