Make the --config and --variable options program-wide
The configuration file only affected the behavior of the test subcommand.
However, we want the configuration file to be able to tune all parts of
Kyua, not just the execution of the tests.  Therefore, the --config and
--variable options need to be made program-wide instead of
subcommand-specific, and this is what this change does.

Doing this is quite invasive: it is now the responsibility of the main
module to load the configuration and pass it onwards to the subcommands
it executes.  However, the previous interface of utils::cmdline made
this impossible, as there was no way to pass an opaque data structure
to the subcommands.  This change resolves this by parametrizing the
base_command class and adding an extra parameter to the main() method
of each subcommand.  This parametrization is what makes this change so
invasive.

Note that, at this moment, the only subcommand that uses configuration
data is test.  Other commands will later be changed to read parameters
from the configuration when needed.

Lastly, note that this comes with no integration tests for the new
global options.  Such tests will be added soon.

This change was r189 in Subversion.