Fix signature of placeholder unmount(2) function
In systems that do not support unmount(2), we define a placeholder function
so that non-conditional code can rely on it.  It turns out that the signature
of this placeholder function was wrong: unmount(2) is supposed to return an
integer instead of void.  Fix this.

Fixes issue 21.

This change was r190 in Subversion.