Add the 'config' subcommand
The 'config' subcommand provides a mechanism to inspect the configuration
of Kyua after evaluation.  Other than helping the user in knowing what
exactly Kyua is seeing in the configuration files, it also allows us to
implement detailed tests about the behavior of the configuration files and
related flags.

Update issue 11
The implementation of the command is complete.  However, we need to address
a usability problem mentioned in the bug: the user needs to be able to
specify the configuration variables in the configuration file in the same
way as printed by this command.

This change was r191 in Subversion.