Simplify the definition of test-suite-specific variables
Allow direct addressing of non-existent members of the global 'test_suites'
table by instantiating them when first requested.  This is to remove the
test_suites_var() function, which had always been a hack due to my lack of
knowledge about metatables, and allows us to use the more natural syntax:

test_suites.<test-suite-name>.<property> = <value>

or:

test_suites["<test-suite-name>"]["<property>"] = <value>

Update issue 11
This exposes the fact that all the properties specific to test-suites are
defined inside a global 'test-suites' table.  The next step would be to
fully qualify properties that relate to Kyua, as the bug suggests by
prefixing them with 'engine'.  Gotta think about this though.

This change was r192 in Subversion.