Enable detection of unused parameters and fix warnings
Remove the -Wno-unused-parameter flag from the developer mode settings.
This was inherited from ATF through the NetBSD makefiles but never thought
about whether we wanted this flag or not.

This change gets rid of this flag and introduces a new macro,
UTILS_PARAM_UNUSED, to clearly denote function parameters that are not
used on purpose.  This macro is used to silence all warnings raised by
the compiler.  As is quite obvious, while I had tried (manually) to
explicitly prefix unused paramters with the "unused_" string, this was
not consistent in all places where this was needed.

Fixes issue 23

This change was r194 in Subversion.