Change the behavior of developer mode
Developer mode used to be a collection of two related concepts: the
detection and enabling of multiple warning-detection flags, and the
enforcement of such flags plus the enforcement of assertions.

This change splits these two responsibilities: warning-detection flags
are now unconditionally used, and developer mode makes such warnings
be enforced.  The behavior of assertions is not changed.

Additionally, developer mode will only be enabled on repository
checkouts, not on formal releases.  The enforcement of warnings is a
constant source of pain for users, and the standard reply of "disable
developer mode" just does not help.

Fixes issue 22

This change was r195 in Subversion.