Skip installed/user configuration during testing
The installed configuration of the user-specific configuration may be bogus,
so it may cause unexpected errors during our test runs.  Disable loading of
this configuration.

Note that for the tests run under Kyua (not the bootstrap tests), we go the
easy route of setting KYUA_CONFDIR to an invalid value and relying on Kyua
to reset HOME during the execution of the tests.  We should actually be
passing --config=none explicitly to all kyua invocations, but that seems
quite error prone.

This change was r198 in Subversion.