Disable the execution of the bootstrap testsuite from within Kyua
The bootstrap testsuite stalls when run from within Kyua in those systems
where /bin/sh is not bash.  The root cause of this issue appears to be a
bug in autotest, so there is not much we can do here.  Therefore, disable
the execution of the testsuite for now until autotest is fixed or we find
a workaround.

Update issue 24
This is just a workaround for the problem described in this issue and we
really need to find a real solution.  However, I want to get 0.2 released
soon and this is a showstopper.

This change was r200 in Subversion.