Explain that autoreconf has to find atf-sh.m4
Add a note saying that it may be necessary to tell autoreconf where the
atf-sh.m4 file is to fix strange syntax errors.

Update issue 19
This change adds a little note to the installation documentation regarding
the problem described in this issue.  I don't expect many users to read
this document though, so I still should find a way to make the errors
clearer.

This change was r202 in Subversion.