Fix build in Ubuntu 10.04.1 LTS
The fixes here are only to silence some warnings raised by the particular
version of the compiler in this distribution and by the build of a code
path that is not necessary in NetBSD nor Mac OS X.

This does not fix runtime problems though.  All tests are broken in this
Linux version because lchmod(3) is not implemented yet it is available at
compile time.  Gotta figure out how to workaround this...

This change was r203 in Subversion.