Detect broken lchmod(3) implementations and workaround them
Ubuntu 10.04.1 LTS provides a lchmod(3) stub that does nothing other than
return an error code unconditionally.  Because it is a stub, our code
compiles fine and gives the impression that will work at runtime, but the
real result is that *all* Kyua tests fail to execute because Kyua reports
all work directory cleanups as failed.

To resolve this, add a configure check to see if lchmod(3) is present and
works and, if not, fallback to chmod(2) with proper logging of a warning
message.

Of course, one of our tests has to be marked as expected failure when
lchmod(3) is missing.  But that's OK.

This change was r204 in Subversion.