Allow inheriting stdout/stderr in child_with_files
The child_with_files fork method blindly attempts to open the provided
files to store its stdout and stderr.  Add support to treat /dev/stdout
and /dev/stderr differently by just inheriting the streams of the parent
process.  This is required to be able to implement a "debug" command
that just prints out the streams to the user.

This change was r205 in Subversion.