Detect if ATF_CHECK_SH is not available and give a clear error otherwise
If atf is not installed in a common location, autoreconf will not find
atf-sh.m4 automatically and the configure script will blow up with a
cryptic message.  Fix this by detecting the existence of the ATF_CHECK_SH
macro and erroring out in a controlled manner.

Fixes issue 19
It turns out detecting the presence of the necessary atf macros was easier
than I thought.  autoreconf will now give a clear message if it cannot
find atf-sh.m4.

This change was r206 in Subversion.