Add a method to optional to allow querying for a value with a default fallback
The new get_default method to the optional class allows the caller to pass in
a value to be used as a default, in cases where the optional object contains
no value.

While doing this, enable the execution of the optional_test in the test suite.
Don't know how this slipped through!

This change was r207 in Subversion.