Implement the 'debug' command
The debug command allows the execution of a specific test program with
more fine-grained control on the runtime engine.

At the moment, this command allows the user to see the stdout and stderr
of the test case, but in the future will allow more functionality like
poking into the work directory of the test or executing a debugger when
the test case fails.

This change was r208 in Subversion.