Ensure that the detected lua provides lua.hpp
It looks like Lua 5.0 does not provide lua.hpp, whereas Lua 5.1 does.  Our
configure checks did accept Lua 5.0 when the pkg-config files were not
found, because we fall back to the lua-config script that does not provide
version information.  As a result, configure was successful but the build
failed half-way through.

Fix this by ensuring that the detected Lua provides the lua.hpp header.

Update issue 20
Status: Fixed
The problem was actually that the configure script was mistakenly picking
up the Lua 5.0 header files and assuming lua.hpp was present.  This change
has modified the checks to ensure that lua.hpp exists (which in turn forces
the user to have to install the 5.1 packages instead of the 5.0 ones).

This change was r211 in Subversion.