Fix test failures caused by different directory ordering
Some of our tests read the contents of directories and were assuming that
the ordering of the reads matched the ordering of the creation of the
files inside such directories.  This is a wrong assumption that led to
spurious test failures under some cases.  Found under Debian sid with ext4.

This change was r213 in Subversion.