Detect the correct value for optind to reset getopt(3) processing
It turns out that the GNU implementation of the getopt(3) functions requires
optind to be set to 0 to reset option processing (instead of the standard 1)
when using GNU-specific extensions.

This fixes crashes in Debian sid as of today of a couple of integration tests
(those that pass --variable foo=bar twice).  Found using Valgrind.

This change was r214 in Subversion.