Use Lutok instead of Lua
The Lutok project contains what used to be our own utils::lua library.
Switch to using the Lutok source code instead of keeping our own copy
of the Lua wrappers.

I intend to use svn externals to bring in a copy of Lutok into the
distribution to ease the installation of Kyua, while providing a
configure switch to decide whether to use the built-in code or to use
any system-wide code of Lutok.  But this will come later.

No functional changes (supposedly!).

This change was r221 in Subversion.