Adapt to the new C-clean header files of Lutok
The Lutok header files have gone through some severe reorganization to
drop all traces of the inclusion of the Lua C header files.  While we
did not use any of the C-level stuff, some of the interfaces we were
calling have had to change to allow this cleanup.

In particular, change the way we access the Lua debug data and also
simplify the way in which we define our Lua bindings: wrap_cxx_function
is no more!

This change was r224 in Subversion.