Fix test: getopt returns an int, not a char
The test that attempts to determine the value to pass to optind to reset
getopt processing was assuming that the return value of getopt is a char
instead of an int.

In a PowerPC (and possibly other architectures), the char type is always
defined as unsigned (which is allowed by the C standard).  This caused
the test to enter an infinite loop because the terminator return value
of getopt (-1) was being converted to 255 and thus never matched the
loop exit condition.

I never saw this in my Mac Mini G4 because I always built the code out
of the repository.  In this case, -Werror is enabled and the test just
failed to compile altogether.  However, if we build Kyua from pkgsrc,
-Werror is disabled and the test enters an infinite loop.

Problem reported and diagnosed by Havard Eidnes in kyua-discuss@.

This change was r230 in Subversion.