Improve the database-opening API to expose less magic
Split the open() constructor in three: in_memory(), open() and temporary().
The former and the latter are used to hide the special cases in which open()
would take the magic constants ":memory:" and "".  Note that it was already
impossible to pass "" to open() because the empty string is an invalid path
and the fs::path class would not have allowed it.

While doing this, also fix double-releasing of resources when the instances
of the database have been copied.  I did not see any problem because G++
implements the return-copy-optimization... but this surely could have caused
problems in other compilers and in my own code if I made explicit copies of
the objects.  Added a test for this case.

This change was r232 in Subversion.