Move the filters module from the cli to the engine
Filtering test cases is an operation that does not belong in the CLI.
(This was actually stated as a todo but not done yet.)

Most of the code that currently resides in the CLI should really be
part of the engine and, in order to be able to move it, the filters
have to be moved first.  I'm pretty sure that parts of
cli::filters_state will have to be moved too, but this will be done
as part of a separate refactoring.

This change was r236 in Subversion.