Move filters_state into the engine
The filters_state class is a generic object used to track the match state
of a collection of filters.  This has no relation to presentation, so it
belongs into the engine.  (In fact, this is a prerequisite to move even
more logic code from the cli layer into the engine layer.)

While doing this, split the presentation aspects of filters_state (parsing
of filters and reporting of unused filters) out of the class and put these
into the cli code as auxiliary functions.

This change was r237 in Subversion.