Move the 'cmd_test' logic to a new 'run_tests' engine driver
Introduce "engine drivers" (aka controllers in the more traditional
terminology) by moving all of the logic involved in executing multiple
tests from the presentation layer (cli/cmd_test) to a driver module
inside the engine (engine/drivers/run_tests).  More similar changes
will follow.

This is still quite clunky in the sense that there are still bits of
presentation left in the engine.  These will be cleaned up later.
Also, there are no tests for the driver module yet.

The goal of this change is to allow the addition of the database for
test results without disturbing front-end code.  The coordination of
execution of tests and storage of results is something that belongs
in the engine.

This change was r238 in Subversion.