Turn result_ptr into a shared_ptr
result_ptr had been an auto_ptr until now.  This causes issues when
attempting to return pointers as part of other structures and, in
all seriousness, result objects should really be able to be passed
around at will.

The good thing is that this change is trivial :-)  No other code needs
to change to cope with this modification.

This change was r240 in Subversion.