Move the 'cmd_debug' logic to a new 'debug_test' engine driver
This follows the previous change that did the same thing for cmd_test
and cmd_list.  The presentation code for cmd_debug is now focused on
the cli layer, without any interferences from runtime logic.
Similarly, the runtime logic is now all encapsulated within the engine.

As happened with the cmd_test change, this comes with no new tests.
In other words, debug_test_test is empty at the moment.  I will surely
find a chance to write such tests later as the features of debug_test
grow.  But for now, this is good enough.

This change was r241 in Subversion.