Use ATF_REQUIRE to compare enum values, not ATF_REQUIRE_EQ
For some reason, GCC 4.5.3 in NetBSD/macppc does not like us to compare
enum values with ATF_REQUIRE_EQ.  The default converesions of the enum
type to an integer are ambiguous, and this breaks the build because the
compiler cannot choose the correct overload for operator<<.  Fix this
by using ATF_REQUIRE to compare the values instead.

This change was r243 in Subversion.