Use symlink(2) instead of link(2)
In r239, I was debugging a test case and changed the code to use link(2)
instead of symlink(2).  This was supposed to be temporary only, but somehow
it made it into the repository.  Revert such change and use symlink(2)
again.

This fixes test failures that I was experiencing on my macppc.  The reason
for the failures is that my /tmp directory is a tmpfs file system, and thus
the test program could not hardlink the helper binary from my home directory
into the work directory (different file systems).

This change was r244 in Subversion.