Implement the last_insert_rowid method in sqlite::database
Allow querying the row identifier of the last insertion at the database
level.  Not sure if this will be enough to get the identifiers of the
newly created entities (it'll work, but probably not when we have
concurrency), nor if there is a better way to do it.  We'll see.

This change was r245 in Subversion.