Remove load_kyuafile
The usefulness of load_kyuafile was removed a long time ago, as its
stale comments indicate.  At the moment, it was only a wrapper around
the actual kyuafile::load function.  Furthermore, the loading of
Kyuafiles as part of specific commands now happens in engine/drivers/.
Therefore, remove an obsolete function.

This change was r248 in Subversion.