Initial implementation of the data store layer (mostly a placeholder)
This change adds a new "layer" to the code base to represent and abstract
the access to the backend database into which we will store test results.

At the moment, the database schema includes just a metadata table to hold
database-wide values.  The current code implements mechanisms to 1) create
a new database and 2) open existing databases while verifying that their
metadata is good.  As mentioned in the subject, this is mostly to have a
placeholder onto which build additional stuff later on.

The new code is not yet hooked up into the tool but is enabled during for
building and testing.

This change was r253 in Subversion.