Add the db-exec subcommand and the --store option
This change adds the db-exec subcommand, which provides a mechanism to run
arbitrary SQL commands within the store database.  This is useful for
debugging purposes, but will also come in handy during the development of
the database.

Also, this adds support for a "generic" --store option, which will later
be used by all the commands that need to talk to the database (similar to
the --kyuafile option).

This change was r255 in Subversion.