Add the step_without_results method to the sqlite::statement
The step_without_results function is supposed to be a wrapper around the
already-existing step() method.  The difference is that the new method
does not return any values: it is intended to be run in any statement that
yields no data, and thus the method ensures that this is true.

This change was r256 in Subversion.