Make column_text return a string
If column_text returns a pointer to a char, it is extremely easy to make
programming mistakes.  For example, something like ATF_REQUIRE_EQ("foo",
stmt.column_text(0)) will produce an invalid result because the macro
is using the equality operator on two pointers.

Therefore, change column_text to return a std::string.  This is to avoid
the pitfalls described above and also to remain more in line with the
rest of the code.  Note that column_blob remains the same (returning a
pointer) because, well, blobs are special entities and they are
potentially very large (which would make copies unfeasible).

This change was r259 in Subversion.