Add actions and contexts
These two new concepts are part of the database design.  Actions are needed
to group the results of all tests of a particular run and to be able to
maintain historical data.  Contexts are related to actions in that they
store a description of the runtime environment of the action.

This change adds two new class definitions in the engine layer to represent
both actions and context.  Additionally, it modifies the store layer to be
able to persist such concepts by adding them to the schema and by adding a
new store::transaction type that abstracts the get/put of objects from/to
the database.  At the moment, this only implements the "put" part of the
backend.  Retrieving objects will come later.

This change was r260 in Subversion.