Make the put methods return the identifier of the inserted object
The caller may be interested in knowing the identifier of the objects it
puts into the database for a variety of purposes, so return it.  For
example, we need a way to show the identifier to the user when he has
run an action, and therefore we need to know what the identifier is.

This change was r262 in Subversion.