Hook the database into the 'test' command
Make 'test' open the data store and generate a new persistent action in it.
No test results are stored yet, as we do not have support to persist test
results at this point.

This also means that the 'test' command now gains the --store flag to specify
where the database lives.

This change was r263 in Subversion.