Link against the store library first, then the engine
The tests in the store fail to build in NetBSD because they linked with the
engine first and then with the store, but the store needs to be able to call
back into the engine to execute the unique_address() methods.

Fix this by reversing the ordering of the libraries to link to.  Somehow I
feel this will cause problems later, but I don't have any better solution.
(Replacing the use of templates with a virtual base class that provides
access to things like unique_address() would solve this though... but I've
been trying to avoid virtual base classes on purpose.)

This change was r269 in Subversion.