Add unique_address() to test cases
Convert the test case classes to use the pimpl idiom and add a
unique_address() method to them.  This is a prerequisite to support
persistence later on.

As a side effect, this makes all fields of atf_iface::test_case private
and provides getters to access them.  Should have done this from the
beginning... sloppiness always comes back to bite.

This change was r271 in Subversion.