Add unique_address() to test case results
Convert the test case result classes to use the pimpl idiom and add a
unique_address() method to them.  This is a prerequisite to support
persistence later on.

As a side effect, this adds some common implementation to the test
case results, such as having a simple representation for good/bad
test cases and their "reason".  This gets us rid of some significant
amount of ugly code.  However, the whole results representation is
pretty nasty as it is now.  I'm thinking that the subclassing was
probably a bad idea from the beginning, for very little gain: all
test cases support results with the exception of "passed", and the
current code structure only makes things harder to use and read.

This change was r272 in Subversion.