Remove the object->id mapping crazyness
When the ability to store actions and contexts was added in r260, it seemed
a good idea to keep track of the database identifiers assigned to such
objects.  The identifier was based on the in-memory address of the objects,
which worked fine when the objects existed for the duration of a transaction.
Things break when the objects are emphemeral.  For example, when implementing
the recording of test case results, each test case result is deleted right
after it is stored, so it's likely that the next result will reuse the same
identifier.  And... kaboom, we detect a duplicate identifier and we crash.

This could easily be fixed by changing the unique_address() methods to return
a sequential, program-wide identifier.  However, this whole idea does not seem
to go in the right direction.  We don't want to be able to detect double-puts
(so that we can update records if needed).  Records are supposed to be
immutable, so we will be executing puts once per object, and that's it.

So we have two alternative approaches:
1. Expose a single put_action method in the transaction that takes a single
   object and stores such object *plus* all of the dependencies in the database
   as a single call.  This is very simple and makes updating the database fast
   (short transactions), but we need to keep "a lot" of data in memory before
   it can be flushed to the database.
2. Have thin put_* methods that work on an object-level and let the caller
   perform the id mappings between dependent objects.  This allows the caller
   to insert stuff into the database "slowly" (as the program runs) so we
   prevent having to keep track of such data at runtime other than the ids.
   (Sure, the open transaction will keep this data as well, but this should be
   smart enough to write stuff to disk if the transaction gets too big.)

I'm choosing 2 for now because it fits better with the current implementation
of everything we have, and I think it might behave better at runtime (no, I
don't have any data to back this up).  Maybe we will have to go to 1 at some
point.

The unique_address() methods are staying for now because I don't know yet how
to implement information retrieval from the database... but I can already
imagine they will prove to be useless.  Meh.

This change was r273 in Subversion.