Remove single '.' components from paths
A path such as 'foo/./bar' is the same as 'foo/bar', but we would not
detect this due to the '.' component.  Strip out '.' components when
they are not the first one so that, when we do a to_absolute() call,
we get more deterministic and unambiguous results.

This change was r274 in Subversion.