Implement recording of test case results
This adds recording of test programs, test cases and test case results to
the database during the execution of a "kyua test" command.  Finally!

This is still pretty much a proof of concept and there is still a lot of
work to do to get this to be decent.  Frankly, the code is starting to
be really messy at places...  But so far it works and this leaves us the
option of implementing an also-preliminary "report" command RSN.

This change was r276 in Subversion.