Add "safe" versions of the column_* methods
These new methods raise an exception when the type of the retrieved cell
does not match the expectations of the call, instead of failing with an
assertion as their column_* counterparts do.

Also, and because these methods are only provided to simplify programming,
these take a column name instead of a column identifier as their single
parameter.

This change was r277 in Subversion.