Add methods to retrieve actions and contexts
The new get_action and get_context methods allow the caller to fetch an
action and a context by its identifer, respectively.  These are simple
objects that I know need to be loaded at once, so this is easy to
implement.  Loading the "contents" of an action (e.g. the list of test
programs that compose it) will be trickier and thus done later.

This change was r278 in Subversion.