Add the get_latest_action method
This calculates what the latest action stored in the database is and returns
it to the caller.  This is just in addition to get_action: the user will be
allowed to specify or not the action to report, so we need these two types
of queries.

This change was r279 in Subversion.