Initial addition of the 'report' command
At this point, this is extremely incomplete and should be considered just
a placeholder.  The only functionality implemented so far is the printing
of the runtime context of a given action.  However, because this touches
too many things, it is better to commit this as is for now, and then build
upon it iteratively.

This change was r282 in Subversion.