Get rid of the unique_address fiasco
The unique_address started as a good idea but it is not workable when
objects can be destroyed and vanish from memory... and this happens all
the time.  (Surely, we could have made it work by forcing the objects
to stay around... but why would we when we have simpler alternatives?)

This change was r283 in Subversion.