Fix the test case results insanity
The previous code was using a class hierarchy to represent test case
results, with the idea being that each result can have its own set of
"metadata".  In practice, however, all test case results had a reason
field attached except for "passed", which did not have it.  The levels
of complexity introduced by this "abstraction" were just insane and
made the code unmanageable.

This change simplifies the representation of test case results by
using a single plain class and removes the "passed" vs. everything else
distinction.  It is just not worth it.

A better idea of all this is in the diffstat:
    45 files changed, 547 insertions(+), 1659 deletions(-)

This change was r287 in Subversion.