Change the way "report" works
Instead of dumping all test case results, ignore passed results and
just print those test cases that deserve attention (i.e. all that
are non-passed).

Also exclude the context from the output by default.  It will be
useful to always include it in HTML reports, but not in console
reports...

This is still pretty much an ad-hoc change, as the reports should be
customizable.  But better to get a useful replacement for atf-report
now than perfection.

This change was r288 in Subversion.