Store and load full definitions of the test programs
Add the ability to the database schema to accurately represent atf and
plain test programs, and change the code to be able to store and load
these.

As a side-effect, the "scan_action" driver can now inspec the full data
of test programs, whereas in the past it only had access to the absolute
path of test programs.

Note that this changes the way the "report" command behaves: it is no
longer able to print the absolute path of the recorded test programs,
because we do not capture this information in some cases.  This should
be addressed later.

This change was r289 in Subversion.