Rework the representation of test programs that fail to list their test cases
Up until now, the base_test_program::load_test_cases method was allowed to
throw an exception if the loading failed (e.g. bogus test program).  Such
exception would later be converted by the runner driver to a fake test case
identifier (foo/bar:__test_program__) that was later reported to the
presentation layer for display purposes.  Other code, such as the scanning
of existing test cases (list command), would have to deal with such broken
test programs on its own.

This approach posed consistency issues across the different components that
have to load test case lists, and also introduced a serious conceptual
problem: we were generating a "test case identifier" that does not
correspond to any existing (in memory) test case.  This gets in the way in
my attempts to remove the "test_case_id" class, for example.  It also makes
code more complicated than it should be, because it has to deal with corner
cases (see how the "list" command had to report broken test programs
separately).

This change redoes all this: the failure to list test cases is something
very specific to ATF test cases.  Therefore, it makes sense to represent
such error condition within the atf_iface code alone, and leave the core
logic unaware of this particular issue.  The way we do this now is by
injecting a fake test case object into the test program, representing that
the listing of test cases failed.  Such test case can be "executed", at
which point it will just print the reason that led to the failure.

As a side-effect, broken test programs now get properly recorded in the
database, because we do not have to special-case this condition any more.

This change was r290 in Subversion.