Move the formatting of test case identifiers to the cli layer
Remove the test_case_id::str method and move the formatting of such
identifiers to the common code of the cli as a new format_test_case_id
function.  This is a preliminary step in preparation for the removal
of the test_case_id class (still to be evaluated though).

This change was r291 in Subversion.