Add dependency on libengine.a to libstore.a
The store module requires the engine module, as we dereference objects from
the engine layer to be able to store them in the database.

This introduces a cyclic dependency between libengine.a and libstore.a, as
they need each other.  Because this cyclic dependency only appears during
link time (with the effective result of listing each library more than
once in the link command), I don't think this will cause any issues and it
fixes the build under NetBSD.

This change was r292 in Subversion.