Get rid of engine::test_case_id
The test_case_id class served its purpose a while ago but it had become
useless and it hides real design problems in the code.  This class was
originally introduced as a way to uniquely identify tests, but doing
that does not make much sense.  In practice, it was mostly used to pass
back data to the UI in a quite ad-hoc and non-extensible manner.

The deletion of this class uncovers another problem with the current
cyclic dependency between test cases and their test programs (the other
major problem being difficulty in storing the data in the database).
So, consider this change as a preliminary step to a refactoring on how
the relation between test cases and test programs is represented.

This change was r293 in Subversion.