Fix tests (and build) in NetBSD
The build broke again due to an attempt to use ATF_REQUIRE_EQ on an
enum value, which gcc 4.5.3 does not want to compile because it cannot
infer the correct integer type for printing.

Also, this fixes a faulty test that would fail depending on whether
core dumps are allowed or not by the system policies.

This change was r294 in Subversion.