Make the debug_test driver return the matched test case as a filter
Remove a previously-introduced hack that returned the matched test
case by debug_test as a string with hardcoded formatting.  Instead,
return a test filter that would match that unique test case and let
the caller format it as appropriate.

I am starting to think that breaking the test_case<->test_program
cyclic dependency is maybe not worth it, and the hack being removed
mentioned this as a possible solution.

This change was r295 in Subversion.