Dump logging information for tests on stderr
The 'kyua' binary should not show logging information until we have had
a chance to process the --log* options offered to the user.  However,
this does not make sense for test cases: test cases should log to stderr
from the very beginning so that useful information is recorded for
postmortem analysis.

Do this by introducing a new logging::set_inmemory() call that, when
explicitly called, tells the logging module to buffer messages until a
call to logging::set_persistency() is made.  This new function is used
only by 'kyua', and by some tests that require it.

This change was r296 in Subversion.