Allow setting the test cases of a test program and searching them by name
Extend the base_test_program class to support explicitly setting the test
cases (instead of loading them from the program itself).  This is needed
when loading the test program from the database, in which case we know
the test cases that are part of it (and that may not match what is on disk).

Also add a find method to be able to locate test cases by their name, which
will come in handy later.  While doing this, introduce a new not_found_error
exception type to denote the errors of the find method.

This change was r298 in Subversion.