Add missing docstrings and fix a bunch of incorrect ones
Turn on extraction of docstrings for anonymous namespaces, local classes,
private members and static entities so that Doxygen warns for even more
invalid docstrings and missing ones.

After turning these on, perform all the necessary docstring adjustments to
make the Doxygen warnings happy.

Phew, this was quite a bit of work!

This change was r299 in Subversion.