Make test_program::absolute_path() really return an absolute path
The previous change introduced some tiny code changes to better reflect the
adjustments to the docstrings.  However, I forgot to run the tests and a
precondition added to the test_program module broke all the tests.  To fix
this, make the test_program::absolute_path() really return the absolute path
to the test program.

This change was r300 in Subversion.