Store and load full test case definitions
Change the store code to record the definition of test cases in enough
detail to permit instantiating them during load time using the test
case types we already have (atf_test_case and plain_test_case).

This is to allow report generation and other functionality to have
access to all the test case data that was involved in the execution of
the test cases.

While this works, it is far from perfect.  The code to iterate through
test cases is a bit fragile and the tests are pretty basic.  However,
this is so complex and has been sitting on my work directory for so
long that I have lost track of what is going on in this particular
change.  So I have cleaned it up a little bit and am submitting so that
refinements can be done iteratively.

This change was r303 in Subversion.