Fix build on NetBSD again
It turns out enum values are not correctly swallowed by utils::format
because they cannot be fed to streams (which I already knew).  Fix this.

I wonder if OS X's compiler could fail on this condition as well so that
the problems didn't magically appear when building in NetBSD...

This change was r304 in Subversion.